+++ sismic/testing/steps.py
@@ -66,7 +66,7 @@ def load_statechart(context, path):
     context._events = []
     context._automatic_execution = True
     context._interpreter.bind(context._events.append)
-    context._interpreter.execute_once()  # init
+    _execute_statechart(context, force_execution=True, execute_once=True)
 
 
 @given('I execute the statechart')